[Chore] relocate files
