+++ src/main/java/com/capstone/favicon/admin/domain/FAQ.java
@@ -1,0 +1,36 @@
+package com.capstone.favicon.admin.domain;
+
+import com.capstone.favicon.user.domain.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "faq")
+public class FAQ {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long faqId;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private FAQCategory category;
+
+    @Column(nullable = false, length = 255)
+    private String question;
+
+    @Column(nullable = false, length = 255)
+    private String answer;
+
+    public enum FAQCategory {
+        기타, 데이터라이선스, 문제해결, 서비스이용, 회원정보관리
+    }
+
+}
