+++ src/main/java/com/capstone/favicon/admin/dto/FAQRequestDto.java
@@ -1,0 +1,13 @@
+package com.capstone.favicon.admin.dto;
+
+import com.capstone.favicon.admin.domain.FAQ.FAQCategory;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FAQRequestDto {
+    private FAQCategory category;
+    private String question;
+    private String answer;
+}
