+++ src/main/java/com/capstone/favicon/admin/dto/NoticeResponseDto.java
@@ -1,0 +1,15 @@
+package com.capstone.favicon.admin.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class NoticeResponseDto {
+    private Long noticeId;
+    private String title;
+    private String content;
+    private String createDate;
+    private int view;
+    private String label;
+}
