+++ src/main/java/com/capstone/favicon/admin/repository/FAQRepository.java
@@ -1,0 +1,9 @@
+package com.capstone.favicon.admin.repository;
+
+import com.capstone.favicon.admin.domain.FAQ;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FAQRepository extends JpaRepository<FAQ, Long> {
+}
