+++ src/main/java/com/capstone/favicon/admin/repository/NoticeRepository.java
@@ -1,0 +1,9 @@
+package com.capstone.favicon.admin.repository;
+
+import com.capstone.favicon.admin.domain.Notice;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
+}
