--- src/main/java/com/capstone/favicon/faq/repository/FAQRepository.java
@@ -1,9 +1,0 @@
-package com.capstone.favicon.faq.repository;
-
-import com.capstone.favicon.faq.domain.FAQ;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FAQRepository extends JpaRepository<FAQ, Long> {
-}
