--- src/main/java/com/capstone/favicon/notice/repository/NoticeRepository.java
@@ -1,9 +1,0 @@
-package com.capstone.favicon.notice.repository;
-
-import com.capstone.favicon.notice.domain.Notice;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
-}
