+++ src/main/java/com/capstone/favicon/user/domain/Answer.java
@@ -1,0 +1,32 @@
+package com.capstone.favicon.user.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Setter
+@Table(name="answer")
+public class Answer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long answerId;
+
+    @ManyToOne
+    @JoinColumn(name = "question_id", nullable = false)
+    private Question question;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    // private String title;
+
+    @Column(columnDefinition = "TEXT")
+    private String content;
+
+    private LocalDate createDate;
+}
