+++ src/main/java/com/capstone/favicon/user/domain/Question.java
@@ -1,0 +1,45 @@
+package com.capstone.favicon.user.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Setter
+@Table(name="question")
+public class Question {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long questionId;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(columnDefinition = "TEXT")
+    private String content;
+
+    private LocalDate createDate;
+    /* private String title;
+    private Integer view = 0;
+
+    @Enumerated(EnumType.STRING)
+    private AnswerStatus answerStatus = AnswerStatus.PENDING;
+
+    @Enumerated(EnumType.STRING)
+    private Category category;
+
+    private String imageUrl;
+
+    public enum AnswerStatus {
+        PENDING, COMPLETED
+    }
+
+    public enum Category {
+        GENERAL, TECHNICAL, POLICY
+    } */
+}
+
