+++ src/main/java/com/capstone/favicon/user/domain/Scrap.java
@@ -1,0 +1,26 @@
+package com.capstone.favicon.user.domain;
+
+import com.capstone.favicon.dataset.domain.Dataset;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name="scrap")
+public class Scrap {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long scrapId;
+
+    @ManyToOne
+    @JoinColumn(name = "dataset_id", nullable = false)
+    private Dataset dataset;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
+
